% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/petersen.r
\docType{data}
\name{petersen}
\alias{petersen}
\title{Simulation of clustering with firm and time effects.}
\format{A data frame with 5000 rows and 4 variables}
\description{
A dataset containing the 500 simulated firms over 10 years.
Originally created by Mitchell Petersen in conjunction with
Petersen (2009) and made available at
\url{http://www.kellogg.northwestern.edu/faculty/petersen/htm/papers/se/test_data.txt}.
See the references for simulation process.
The variables are as follows:
}
\details{
\itemize{
  \item firmid.  Firm identifier.
  \item year.  Year identifier.
  \item x. Independent (right-hand side) variable.
  \item y. Dependent (left-hand side) variable.
}
}
\references{
Petersen, M. A. (2009). Estimating standard errors in finance panel data sets: Comparing approaches. Review of financial studies, 22(1), 435-480.

Mitchell Petersen's description of the simulation process: \url{http://www.kellogg.northwestern.edu/faculty/petersen/htm/papers/se/se_programming.htm}
}
\keyword{datasets}

